Change location of CharityAddresses and CharityItemTypes controllers
* Add fixes to email sending logic
